Refactor history service config
