+++ service/history/historyCache.go
@@ -40,6 +40,7 @@ import (
 	"go.temporal.io/server/common/log/tag"
 	"go.temporal.io/server/common/metrics"
 	"go.temporal.io/server/common/persistence"
+	"go.temporal.io/server/service/history/configs"
 )
 
 type (
@@ -52,7 +53,7 @@ type (
 		disabled         bool
 		logger           log.Logger
 		metricsClient    metrics.Client
-		config           *Config
+		config           *configs.Config
 	}
 )
