+++ service/history/shardContextTest.go
@@ -33,6 +33,7 @@ import (
 	"go.temporal.io/server/common/metrics"
 	"go.temporal.io/server/common/persistence"
 	"go.temporal.io/server/common/resource"
+	"go.temporal.io/server/service/history/configs"
 	"go.temporal.io/server/service/history/events"
 )
 
@@ -48,7 +49,7 @@ var _ ShardContext = (*shardContextTest)(nil)
 func newTestShardContext(
 	ctrl *gomock.Controller,
 	shardInfo *persistence.ShardInfoWithFailover,
-	config *Config,
+	config *configs.Config,
 ) *shardContextTest {
 	resource := resource.NewTest(ctrl, metrics.History)
 	eventsCache := events.NewMockCache(ctrl)
