+++ service/history/timerQueueProcessorBase.go
@@ -33,6 +33,7 @@ import (
 	enumsspb "go.temporal.io/server/api/enums/v1"
 	persistencespb "go.temporal.io/server/api/persistence/v1"
 	"go.temporal.io/server/common/primitives/timestamp"
+	"go.temporal.io/server/service/history/configs"
 
 	"go.temporal.io/server/common"
 	"go.temporal.io/server/common/backoff"
@@ -63,7 +64,7 @@ type (
 		status               int32
 		shutdownWG           sync.WaitGroup
 		shutdownCh           chan struct{}
-		config               *Config
+		config               *configs.Config
 		logger               log.Logger
 		metricsClient        metrics.Client
 		metricsScope         metrics.Scope
