+++ service/history/timerQueueTaskExecutorBase.go
@@ -38,6 +38,7 @@ import (
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/metrics"
 	"go.temporal.io/server/common/persistence"
+	"go.temporal.io/server/service/history/configs"
 	"go.temporal.io/server/service/worker/archiver"
 )
 
@@ -48,7 +49,7 @@ type (
 		cache          *historyCache
 		logger         log.Logger
 		metricsClient  metrics.Client
-		config         *Config
+		config         *configs.Config
 	}
 )
 
@@ -57,7 +58,7 @@ func newTimerQueueTaskExecutorBase(
 	historyService *historyEngineImpl,
 	logger log.Logger,
 	metricsClient metrics.Client,
-	config *Config,
+	config *configs.Config,
 ) *timerQueueTaskExecutorBase {
 	return &timerQueueTaskExecutorBase{
 		shard:          shard,
